Draft of WASI support
This is a first pass at implementing support for [WASI](https://wasi.dev/).  It
adds a new `TeaVMTargetType` of `WASI`, which is mostly the same as
`WEBASSEMBLY` except it generates a module suitable for execution by a
WASI-compatible runtime such as
[Wasmtime](https://github.com/bytecodealliance/wasmtime).

Currently supported features:

- A simple allocator for supporting the component model [canonical
  ABI](https://github.com/WebAssembly/component-model/blob/main/design/mvp/CanonicalABI.md)
  calls
- CLI argument marshaling
- Standard in/out/err
- Filesystem operations and I/O
- System clock

See tests/wasi/src/main/java/wasi/Test.java and tests/wasi/test.sh for examples.

Not yet implemented:

- Networking (i.e. sockets)
- Environment variables
- System random numbers
- Process exit, yield, etc.
- [DWARF](https://dwarfstd.org/) debugging support
- Better logging of uncaught exceptions

Note that the current implementation is probably overly paranoid in that all
pointer arguments passed to WASI functions are allocated by the aforementioned
allocator rather than on the Java heap.  I've done this because I'm not clear on
when TeaVM guarantees that object address will not change.  Thus I've assumed
that they could change at any time whatsoever, which is probably not true.

Also note that the filesystem implementation currently throws generic
`IOException`s and `ErrnoException`s rather than those specified by the standard
(e.g. `FileNotFoundException`).  Additional work will be needed to throw the
correct exception based on the `errno` value.

The main goal of this PR is to start a discussion about WASI support in TeaVM
and answer questions such as:

- Is WASI support a reasonable goal for the TeaVM project?
- If so, is the approach of this PR roughly correct, or are significant changes needed?
- What other features are missing to make this useful?

Signed-off-by: Joel Dice <joel.dice@fermyon.com>