revert earlier stub implementations
A previous commit provided stub implementations of `getNativeOffset`,
`toUpperCase`, and `toLowerCase`.  This reverts that change and instead uses the
JS versions in the browser.  It won't work in a WASI runtime without providing
the extra imports to the module, but there's currently no way to get the local
timezone in WASI, and I'm not yet ready to take on a proper implementation of
`toUpperCase` and `toLowerCase`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>